+++ R/get_org_attributes.R
@@ -263,4 +263,3 @@ get_milestone_url <- function(owner, repo, milestone_name) {
   milestone$html_url
 }
 
-