+++ NEWS.md
@@ -8,7 +8,11 @@
 
 -   removes empty milestones in `get_open_milestone_objects()` and `get_open_milestone_object_names()`
 
--   `check_if_updates_since_init()` function 
+-   `check_if_updates_since_init()` function
+ 
+## Minor improvements and bug fixes
+
+-   Adds additional status check to prevent issue creation in `ghqc_create_app()` if there is already an existing issue name of the selected file in the same milestone name.
 
 # ghqc 0.0.0.9003
 
@@ -37,4 +41,3 @@
 ## Minor improvements and bug fixes
 
 -   Fixes `pmx_list()` so that it works on older versions of R.
-